requirements.txt
anyascii==0.3.2
appnope==0.1.3
asgiref==3.7.2
asttokens==2.3.0
backcall==0.2.0
beautifulsoup4==4.11.2
boto3==1.28.39
botocore==1.31.39
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.54.0
django-cogwheels==0.3
django-comments-xtd==2.9.10
django-contrib-comments==2.2.0
django-countries-plus==2.1.0
django-extensions==3.2.3
django-filter==22.1
django-htmx==1.16.0
django-languages-plus==2.0.0
django-modelcluster==6.0
django-permissionedforms==0.1
django-recaptcha==3.0.0
django-storages==1.13.2
django-taggit==3.1.0
django-treebeard==4.7
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
docutils==0.20.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
filetype==1.2.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
ipython==8.15.0
jedi==0.19.0
jmespath==1.0.1
l18n==2021.3
matplotlib-inline==0.1.6
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
polib==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
sentry-sdk==1.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
telepath==0.3.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.26.16
wagtail==5.0.2
wagtail-django-recaptcha==1.0
wagtail-localize==1.5.1
wagtailmenus==3.1.8
wcwidth==0.2.6
webencodings==0.5.1
Willow==1.5.2
